Update README to include the command 'cluster-config --prometheus-exporters-enable'.
Updated the config.py, put 'fallback_scrape_protocol: PrometheusProto' under job names. This fallback protocal is used if a scrape returns blank, unparseable, or invalid Content-Type. Added sort_keys=False to the config file to keep the structure of the code for the prometheus.yml file generation. Without sort keys it would alphabetize the the jobs in the yaml.